+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.3.4] - 2018-6-9
+### Fixed
+- Export module Differencify
+- Make Differencify running with vanila node
+- Fix screenshot path issue for non jest
+
 ## [1.3.3] - 2018-6-6
 ### Fixed
 - Bug fixed for custom test path